Moved files to /plugin
